+++ plugin/src/main/kotlin/de/infolektuell/gradle/jextract/extensions/JextractExtension.kt
@@ -4,9 +4,7 @@ import org.gradle.api.Action
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.model.ObjectFactory
-import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Nested
-import org.gradle.api.tasks.SourceSet
 import javax.inject.Inject
 
 abstract class JextractExtension @Inject constructor(objects: ObjectFactory) {
@@ -25,8 +23,6 @@ abstract class JextractExtension @Inject constructor(objects: ObjectFactory) {
     }
     /** specify the directory to place generated files of all libraries by default */
     abstract val output: DirectoryProperty
-    /** The generated Java sources will be added to this source set (main by default) */
-    abstract val sourceSet: Property<SourceSet>
     companion object {
         const val EXTENSION_NAME = "jextract"
     }